Improve init script

+++ init.sh
@@ -1,28 +1,40 @@
 #!/usr/bin/env bash
+# source init.sh to initialize the env
 
-KAFKA_URL="https://archive.apache.org/dist/kafka/3.2.3/kafka_2.13-3.2.3.tgz"
+KAFKA_VERSION="3.2.3"
 STRIMZI_IMAGE="registry.redhat.io/amq7/amq-streams-kafka-32-rhel8:2.2.0"
 NAMESPACE="test"
 
 echo "Checking prerequisites"
-if [[ -z $OCP_API_URL || -z $OCP_ADMIN_USR || -z $OCP_ADMIN_PWD ]]; then
-  echo "Missing OpenShift parameters" && return 1
-  exit 1
-fi
 for x in curl oc kubectl openssl keytool unzip yq jq git java javac jshell mvn; do
   if ! command -v "$x" &>/dev/null; then
     echo "Missing required utility: $x" && return 1
-    exit 1
   fi
 done
 
-echo "Getting Kafka from $KAFKA_URL"
+get_kafka() {
+  local home && home="$(find /tmp -name 'kafka.*' -printf '%T@ %p\n' 2>/dev/null |sort -n |tail -n1 |awk '{print $2}')"
+  if [[ -n $home ]]; then
+    local version && version="$("$home"/bin/kafka-topics.sh --version 2>/dev/null |awk '{print $1}')"
+    if [[ $version == "$KAFKA_VERSION" ]]; then
+      echo "Getting Kafka from /tmp"
+      KAFKA_HOME="$home" && export KAFKA_HOME \
+        && PATH="$KAFKA_HOME/bin:$PATH" && export PATH
+      PATH="$KAFKA_HOME/bin:$PATH" && export PATH
+      return
+    fi
+  fi
+  echo "Getting Kafka from ASF"
+  KAFKA_HOME="$(mktemp -d -t kafka.XXXXXXX)" && export KAFKA_HOME \
+    && PATH="$KAFKA_HOME/bin:$PATH" && export PATH
+  curl -sLk "https://archive.apache.org/dist/kafka/$KAFKA_VERSION/kafka_2.13-$KAFKA_VERSION.tgz" \
+    | tar xz -C "$KAFKA_HOME" --strip-components 1
+}
+
 pkill -f "kafka.Kafka" ||true
 pkill -f "quorum.QuorumPeerMain" ||true
 rm -rf /tmp/kafka-logs /tmp/zookeeper
-KAFKA_HOME="$(mktemp -d -t kafka.XXXXXXX)" && export KAFKA_HOME
-curl -sLk "$KAFKA_URL" | tar xz -C "$KAFKA_HOME" --strip-components 1
-PATH="$KAFKA_HOME/bin:$PATH" && export PATH
+get_kafka
 
 # find coordinating partition (JDK 9+ required)
 find_cp() {
@@ -31,24 +43,45 @@ find_cp() {
   if [[ -n $id && -n $part ]]; then
     echo 'public void run(String id, int part) { System.out.println(abs(id.hashCode()) % part); }
       private int abs(int n) { return (n == Integer.MIN_VALUE) ? 0 : Math.abs(n); }
-      run("'"$id"'", '"$part"');' \
-      | jshell -
+      run("'"$id"'", '"$part"');' | jshell -
   fi
 }
 
-echo "Authenticating to $OCP_API_URL"
-oc login -u "$OCP_ADMIN_USR" -p "$OCP_ADMIN_PWD" "$OCP_API_URL" --insecure-skip-tls-verify=true &>/dev/null
-kubectl config set-context --current --namespace=$NAMESPACE &>/dev/null
+authn_ocp() {
+  echo "Authenticating to OpenShift"
+  local login="/tmp/ocp-login"
+  if [[ ! -f $login ]]; then
+    read -rp "API URL: " OCP_URL
+    read -rp "Username: " OCP_USR
+    read -rp "Password: " -s OCP_PWD && echo ""
+    declare -px OCP_URL OCP_USR OCP_PWD > "$OCP_LOGIN"
+  else
+    # shellcheck source=/dev/null
+    source "$login"
+    if [[ -z $OCP_URL || -z $OCP_USR || -z $OCP_PWD ]]; then
+      echo "Missing OpenShift parameters" && return 1
+    fi
+  fi
+  if oc login -u "$OCP_USR" -p "$OCP_PWD" "$OCP_URL" --insecure-skip-tls-verify=true &>/dev/null; then
+    return
+  else
+    echo "Authentication failed"
+    rm -rf $login
+    return 1
+  fi
+}
 
-echo "Deploying cluster-wide operators"
-kubectl delete ns $NAMESPACE &>/dev/null ||true
-kubectl -n openshift-operators delete csv --all &>/dev/null ||true
-kubectl -n openshift-operators delete sub --all &>/dev/null ||true
-kubectl create ns $NAMESPACE
-kubectl create -f sub.yaml
+if authn_ocp; then
+  kubectl config set-context --current --namespace="$NAMESPACE" &>/dev/null
+  kubectl delete ns "$NAMESPACE" &>/dev/null ||true
+  kubectl create ns "$NAMESPACE"
+  kubectl -n openshift-operators delete csv --all &>/dev/null ||true
+  kubectl -n openshift-operators delete sub --all &>/dev/null ||true
+  kubectl create -f sub.yaml
 
-krun_kafka() { kubectl run krun-"$(date +%s)" -it --rm --restart="Never" --image="$STRIMZI_IMAGE" -- "$@"; }
+  krun_kafka() { kubectl run krun-"$(date +%s)" -it --rm --restart="Never" --image="$STRIMZI_IMAGE" -- "$@"; }
 
-echo "Environment READY!"
-echo "  |__Kafka home: $KAFKA_HOME"
-echo "  |__Current namespace: $NAMESPACE"
+  echo "Environment READY!
+    |__ Kafka home: $KAFKA_HOME
+    |__ Current namespace: $NAMESPACE"
+fi
